pyproject.toml
[tool.poetry]
name = "openblog"
version = "0.1.0"
description = "A production-ready developer blog platform"
authors = ["Sebastian Gomez"]
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.1"
flask-sqlalchemy = "^3.1"
SQLAlchemy = "^2.0"
alembic = "^1.14"
psycopg2-binary = "^2.9"
redis = "^5.2"
celery = "^5.4"
PyJWT = "^2.10"
Authlib = "^1.4"
bleach = "^6.2"
Markdown = "^3.7"
Pygments = "^2.18"
argon2-cffi = "^23.1"
Flask-WTF = "^1.2"
Flask-Limiter = {version = "^4.0", extras = ["redis"]}
python-dotenv = "^1.0"
gunicorn = "^25.1.0"
prometheus-flask-exporter = ">=0.23"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9"
pytest = "^8.3"
pytest-flask = "^1.3"
factory-boy = "^3.3"
pre-commit = "^4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: requires external services (PostgreSQL + Redis)",
]
